Fix exit model in wallet
